refractor: move Instructions.js and Howto.js to their own subfolder
